move to shared: intent-detector
