+++ client/cody/src/chat/recipes/chat-question.ts
@@ -1,3 +1,4 @@
+import { IntentDetector } from '@sourcegraph/cody-shared/src/intent-detector'
 import { MAX_CURRENT_FILE_TOKENS, MAX_HUMAN_INPUT_TOKENS } from '@sourcegraph/cody-shared/src/prompt/constants'
 import { populateCodeContextTemplate } from '@sourcegraph/cody-shared/src/prompt/templates'
 import { truncateText } from '@sourcegraph/cody-shared/src/prompt/truncation'
@@ -6,7 +7,6 @@ import { getShortTimestamp } from '@sourcegraph/cody-shared/src/timestamp'
 import { CodebaseContext } from '../../codebase-context'
 import { ContextMessage, getContextMessageWithResponse } from '../../codebase-context/messages'
 import { Editor } from '../../editor'
-import { IntentDetector } from '../../intent-detector'
 import { renderMarkdown } from '../markdown'
 import { Interaction } from '../transcript/interaction'
