+++ client/cody/src/chat/recipes/git-log.ts
@@ -1,13 +1,13 @@
 import { spawnSync } from 'child_process'
 import * as path from 'path'
 
+import { IntentDetector } from '@sourcegraph/cody-shared/src/intent-detector'
 import { MAX_RECIPE_INPUT_TOKENS } from '@sourcegraph/cody-shared/src/prompt/constants'
 import { truncateText } from '@sourcegraph/cody-shared/src/prompt/truncation'
 import { getShortTimestamp } from '@sourcegraph/cody-shared/src/timestamp'
 
 import { CodebaseContext } from '../../codebase-context'
 import { Editor } from '../../editor'
-import { IntentDetector } from '../../intent-detector'
 import { renderMarkdown } from '../markdown'
 import { Interaction } from '../transcript/interaction'
